+++ src/include/Timer.hpp
@@ -18,8 +18,8 @@ public:
   unsigned long tick();
   unsigned long tock();
   unsigned long delta() const;
-  unsigned long keyFrameRate() const;
-  unsigned long keyFrameRate( unsigned long newFrameRate );
+  float keyFrameRate() const;
+  float keyFrameRate( float newFrameRate );
   double scale() const;
 
 private:
@@ -32,7 +32,7 @@ private:
 #endif
   unsigned long _delta;
   double _scale;
-  unsigned long _keyFrameRate;
+  float _keyFrameRate;
 };
 
 /**
